+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Improve clarity of LaTeX and BibTeX error log processing.
+- Corrected NameSpace attribute in `reprepbuild.scripts.manifest`.
 
 
 ## [1.7.1] - 2023-12-01